Add isNull method to Pointer class
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@3146 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ rexxapi/server/platform/windows/APIService.cpp
@@ -353,7 +353,7 @@ BOOL Install()
                   0,
                   REG_EXPAND_SZ,
                   (CONST BYTE*)szFilePath,
-                  strlen(szFilePath) + 1);
+                  (DWORD)strlen(szFilePath) + 1);
 
     // Set the supported types flags.
     dwData = EVENTLOG_ERROR_TYPE | EVENTLOG_WARNING_TYPE | EVENTLOG_INFORMATION_TYPE;
